Initial checkin
This contains the program to predict whether a child is selected or rejected based on the two exams score. its a logistic/classification problem.

Copy the data files "*.txt" to your local drive. Open the Matlab file and copy the code from it. Paste it in octave or Matlab and run it. Make sure your working directory contains data files that you copied from here.